Moved some components under common
