+++ src/common/components/relational-information-block/relation-information-block.styles.tsx
